Generic Logger implementation
This PR introduces the Logger. A Logger is a struct that contains a Log(Severity, payload) function. The loggers that are provided are:
- Standard Logger
- Contextual Logger
- Multi Logger
- Google Logger

The standard logger will log to the standard output, attempting to json marshal the payload. If JSON amrshaling doesn't work, it prints the go object version of the output.

The google logger will log to the google cloud console.

The contextual logger takes in another logger and adds a context object to the payload. If the payload satisfies the ContextualEntry interface, it will use the interface to set the log context. Otherwise, it will return a StandardContextualEntry object with the log message and the context as separate fields.

The multi logger logs to multiple loggers at the same time.

Logs such as the standard/google log should check if the log entry implements the Defaults interface, which allows the code to set a default in the log entry itself. For example, the MetricEntry struct implements this interface and sets the "Type" to metric.

In addition, Google & Standard logger implement a blacklist by severity type.

MetricEntry
===========

A metric entry is a log which presents in a specific way to the logging platform. One can query this specific log structure to retrieve a list of similar logs & aggregate them for metrics.

HTTPRequest
==========

HTTP request loggers introduec further functionality to logging, by allowing the client to set to the log context
- whether or not the response was cached
- response status code
- latency

This is used for the Google logger's BaseEntry. See GoogleHTTPRequest logger for more details. Note that if *logging.HTTPRequest is to be used as the context in a Contextual logger, it should be wrapped using `NewHTTPRequestContext` which implements JSON marshalling properly.