Move cancel function into the close for io.ReadCloser for GCP
GCP storage was causing issues in that the context was cancelling before the reader was fully read. To fix this, we wrap the cancel function for the context into the ReadCloser throuhg a special struct. This allows us to have a timeout for the byte reading operation while also retaining the same return signature.